search strategy can use randomness
This is a big improvement: unit tests will more reliably be simplified
due to the redefinition of runStrategy.

Besides that, doing a breadth-first search of the complications, as was
done before, was a silly idea anyway. It's better to have the strategy
pseudorandomly pick one potentially much more complicated point. This
way it's less likely that a search would be run multiple times at the
same point.

It should be possible to property test search strategy w.r.t. a given
ordering.