improve test pretty printer
The things needed to reproduce the failure are now all grouped together
in order: dynamic input, static input, and ouptut.

The domain part is not strictly necessary: you don't need to search the
whole domain again. But it's still informative. However, the domain used
in a composite test may depend upon IO, and may be defined inline. In
general, a domain can't really be identified.