remove notion of static/dynamic input parts
There is now just one input to a test. The distinction between static and
dynamic input can be made when defining a domain. Checking a proprety test
is analogous to applying a test to a domain, which can be done in various
ways depending upon how many samples you want to take and which seed to
use.