remove Declaration monad, simplify Composite
This is much easier to understand. A Composite may check any test which
is in a static pointer and has been given a name, renderer and optional
source location.

The canonical way of making such a thing is to use

  declare name renderer (static test)

for some test. But these declarations may be made anywhere.